+++ 12 - Herencia/EX_4/src/Entidades/Staff.java
@@ -21,7 +21,7 @@ public class Staff extends Employee {
     public Staff() {
     }
 
-    public Staff(Sections sec, Integer yearEnter, Integer office, String completeName, Integer id, MaritalStatus est, Integer age, Sex s, Integer cellNum, String adress, Integer adressNum, String nationality) {
+    public Staff(Sections sec, Integer yearEnter, Integer office, String completeName, Integer id, MaritalStatus est, Integer age, Sex s, Long cellNum, String adress, Integer adressNum, String nationality) {
         super(yearEnter, office, completeName, id, est, age, s, cellNum, adress, adressNum, nationality);
         this.sec = sec;
     }
